+++ src/main/java/li/cil/tis3d/api/module/traits/ModuleWithBakedModel.java
@@ -1,14 +1,19 @@
 package li.cil.tis3d.api.module.traits;
 
 import net.minecraft.block.BlockState;
+import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.model.BakedQuad;
 import net.minecraft.util.Direction;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.IBlockDisplayReader;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
 import net.minecraftforge.client.model.data.IModelData;
 
 import javax.annotation.Nullable;
+import java.util.Collections;
 import java.util.List;
+import java.util.OptionalInt;
 import java.util.Random;
 
 /**
@@ -22,9 +27,55 @@ import java.util.Random;
  */
 public interface ModuleWithBakedModel {
     /**
-     * Called to obtain quads to use for the specified side instead of the casing's default ones. May return
-     * <c>null</c> to not override the default quads. Will be called directly from the casing's
-     * {@link net.minecraft.client.renderer.model.IBakedModel#getQuads(BlockState, Direction, Random, IModelData)}
+     * Whether this module has a module to render.
+     * <p>
+     * If this returns {@code false}, none of the other methods will be called, and the regular
+     * blank module face will be rendered.
+     * <p>
+     * This supports dynamically changing models, which is used by the facade module, for example.
+     *
+     * @return whether this module has a model.
+     */
+    default boolean hasModel() {
+        return false;
+    }
+
+    /**
+     * Collect model data that is needed to render the quads returned by this module.
+     * <p>
+     * The returned value will be passed back into {@link #getQuads(BlockState, Direction, Random, IModelData)}.
+     *
+     * @param world the render-thread safe world access.
+     * @param pos   the position of the casing.
+     * @param state the current block-state of the casing.
+     * @param data  the incoming model data to wrap/expand.
+     * @return model data needed for rendering.
+     */
+    @OnlyIn(Dist.CLIENT)
+    default IModelData getModelData(final IBlockDisplayReader world, final BlockPos pos, final BlockState state, final IModelData data) {
+        return data;
+    }
+
+    /**
+     * Called to obtain quads to use for the specified side instead of the casing's default ones. Will be called
+     * directly from the casing's {@link net.minecraft.client.renderer.model.IBakedModel#getQuads(BlockState, Direction, Random, IModelData)}
+     * logic.
+     *
+     * @param state  the casing's block state.
+     * @param side   the side to obtain replacement quads for.
+     * @param random the random seed to use for the quad generation.
+     * @return the list of replacement quads, or <c>null</c> to use the default casing quads.
+     * @deprecated Use the overload {@link #getQuads(BlockState, Direction, Random, IModelData)} with more context instead.
+     */
+    @OnlyIn(Dist.CLIENT)
+    @Deprecated
+    default List<BakedQuad> getQuads(final @Nullable BlockState state, @Nullable final Direction side, final Random random) {
+        return Collections.emptyList();
+    }
+
+    /**
+     * Called to obtain quads to use for the specified side instead of the casing's default ones. Will be called
+     * directly from the casing's {@link net.minecraft.client.renderer.model.IBakedModel#getQuads(BlockState, Direction, Random, IModelData)}
      * logic.
      *
      * @param state  the casing's block state.
@@ -33,5 +84,35 @@ public interface ModuleWithBakedModel {
      * @return the list of replacement quads, or <c>null</c> to use the default casing quads.
      */
     @OnlyIn(Dist.CLIENT)
-    List<BakedQuad> getQuads(final @Nullable BlockState state, @Nullable final Direction side, final Random random);
+    default List<BakedQuad> getQuads(final @Nullable BlockState state, @Nullable final Direction side, final Random random, final IModelData data) {
+        return getQuads(state, side, random);
+    }
+
+    /**
+     * Returns whether the quads returned by {@link #getQuads(BlockState, Direction, Random, IModelData)} can be
+     * rendered in the specified render layer.
+     *
+     * @param layer the render layer to test.
+     * @return {@code true} if the quads can be rendered in the specified layer; {@code false} otherwise.
+     */
+    @OnlyIn(Dist.CLIENT)
+    default boolean canRenderInLayer(@Nullable final RenderType layer) {
+        return false;
+    }
+
+    /**
+     * Get the tint color to use for the quads returned by this module.
+     * <p>
+     * Note that only the first tint color by any module in one casing will be used. Using facades with varying tint
+     * color in a single casing will lead to wrong results. There's no way around it, so we just live with it.
+     *
+     * @param level     the render-thread safe world access.
+     * @param pos       the position of the casing.
+     * @param tintIndex the tint index to resolve.
+     * @return the color for the specified tint index, if possible.
+     */
+    @OnlyIn(Dist.CLIENT)
+    default OptionalInt getTintColor(@Nullable final IBlockDisplayReader level, @Nullable final BlockPos pos, final int tintIndex) {
+        return OptionalInt.empty();
+    }
 }